+++ evaluator/Social/social_evaluator.py
@@ -22,9 +22,8 @@ class SocialEvaluator(AbstractEvaluator, threading.Thread):
         self.load_config()
 
     @classmethod
-    def get_config_file_name(cls):
-        return "{0}/{1}/{2}/{3}/{4}".format(TENTACLES_PATH, TENTACLES_EVALUATOR_PATH, CONFIG_EVALUATOR_SOCIAL
-                                            , EVALUATOR_CONFIG_FOLDER, cls.get_name() + CONFIG_FILE_EXT)
+    def get_config_file_name(cls, config_evaluator_type=CONFIG_EVALUATOR_SOCIAL):
+        return super().get_config_file_name(config_evaluator_type)
 
     def stop(self):
         self.keep_running = False